feat: add Bohrium environment variable support and optimize file download functionality
